Create README.md
Added README with description of the Portals Island project. This Unreal Engine prototype is built mostly in C++ and includes a mirror portal, a procedural Perlin noise terrain, and an island with linked portals and a gun system. Features like projectiles, dynamic color splashes, and Niagara FX are all driven through programming to highlight gameplay systems and technical design.